Initial commit
This module provides a way of managing playerdata without having to
worry about loading/saving any data. Data is bound to modules, however
it can be "cross-accessed" if the module name is known. If a module
changes its name, the migrate function can be used to update data
ownership accordingly.

Requires APIv3.2 or higher